test(testing): Settings и AppSettings tests вынесены в отдельные suite files
persistence, service, theme, screen controller, presentation и state tests для settings flow выделены из RSSReaderTests.swift в самостоятельные feature-oriented suites
